+++ src/gloo/IBO.h
@@ -8,8 +8,8 @@
 
 class IBO {
 public:
-	IBO(const size_t t_size, const float* t_data);
-	IBO(const std::vector<float>& t_data);
+	IBO(const size_t t_size, const unsigned int* t_data);
+	IBO(const std::vector<unsigned int>& t_data);
 	~IBO();
 
 	void bind() const;
